% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satann.R
\name{prepare_for_SaTAnn}
\alias{prepare_for_SaTAnn}
\title{Prepare the "for_SaTAnn" file}
\usage{
prepare_for_SaTAnn(annotation_file, bam_file,
  path_to_rl_cutoff_file = NA, chunk_size = 5e+06,
  path_to_P_sites_plus_bw = NA, path_to_P_sites_minus_bw = NA,
  path_to_P_sites_uniq_plus_bw = NA,
  path_to_P_sites_uniq_minus_bw = NA,
  path_to_P_sites_uniq_mm_plus_bw = NA,
  path_to_P_sites_uniq_mm_minus_bw = NA, dest_name = NA)
}
\arguments{
\item{annotation_file}{Full path to the annotation file (*Rannot)}

\item{bam_file}{Full path to the bam file}

\item{path_to_rl_cutoff_file}{path to the rl_cutoff_file file specifying in 3 columns the read lengths, cutoffs and compartments ("nucl" for standard chromosomes)}

\item{chunk_size}{the number of alignments to read at each iteration, defaults to 5000000, increase when more RAM is available}

\item{path_to_P_sites_plus_bw}{path to a bigwig file containing P_sites positions on the plus strand}

\item{path_to_P_sites_minus_bw}{path to a bigwig file containing P_sites positions on the minus strand}

\item{path_to_P_sites_uniq_plus_bw}{(Optional) path to a bigwig file containing uniquely mapping P_sites positions on the plus strand}

\item{path_to_P_sites_uniq_minus_bw}{(Optional) path to a bigwig file containing uniquely mapping P_sites positions on the minus strand}

\item{path_to_P_sites_uniq_mm_plus_bw}{(Optional) path to a bigwig file containing uniquely mapping (with mismatches) P_sites positions on the plus strand}

\item{path_to_P_sites_uniq_mm_minus_bw}{(Optional) path to a bigwig file containing uniquely mapping (with mismatches) P_sites positions on the minus strand}

\item{dest_name}{prefix to use for the output files. Defaults to same as \code{bam_file} (appends "for_SaTAnn" to its filename)}
}
\description{
Prepare the "for_SaTAnn" file
}
\details{
This function uses a list of pre-determined read lengths, cutoffs and compartments to calculate P_sites positions.\cr
Alternatively, bigwig files containing P_sites position for each strand can be specified. Optional bigwig files for uniquely mapping P_sites position (with and without mismatches)
can be specified to obtain more statistics on the SaTAnn-identified ORFs
}
\seealso{
\code{\link{run_SaTAnn}}
}
\author{
Lorenzo Calviello, \email{calviello.l.bio@gmail.com}
}
\keyword{SaTAnn}
