% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satann.R
\name{run_SaTAnn}
\alias{run_SaTAnn}
\title{Run the SaTAnn pipeline}
\usage{
run_SaTAnn(for_SaTAnn_file, annotation_file, n_cores,
  prefix = for_SaTAnn_file, gene_name = NA, gene_id = NA,
  genomic_region = NA, write_temp_files = T, write_GTF_file = T,
  write_protein_fasta = T, interactive = T,
  stn.orf_find.all_starts = T, stn.orf_find.nostarts = F,
  stn.orf_find.start_sel_cutoff = NA,
  stn.orf_find.start_sel_cutoff_ave = 0.5,
  stn.orf_find.cutoff_fr_ave = 0.5, stn.orf_quant.cutoff_cums = NA,
  stn.orf_quant.cutoff_pct = 2, stn.orf_quant.cutoff_P_sites = NA)
}
\arguments{
\item{for_SaTAnn_file}{REQUIRED - path to the "for_SaTAnn" file containing P_sites positions and junction reads}

\item{annotation_file}{REQUIRED - path to the *Rannot R file in the annotation directory used in the \code{prepare_annotation_files function}}

\item{n_cores}{REQUIRED - number of cores to use}

\item{prefix}{prefix to use for the output files. Defaults to same as \code{for_SaTAnn_file} (appends to its filename)}

\item{gene_name}{\code{character} vector of gene names to analyze.}

\item{gene_id}{\code{character} vector of gene ids to analyze}

\item{genomic_region}{\code{GRanges} object with genomic regions to analyze}

\item{write_temp_files}{write temporary files. Defaults to \code{TRUE}}

\item{write_GTF_file}{write a GTF files with the ORF coordinates. Defaults to \code{TRUE}}

\item{write_protein_fasta}{write a protein fasta file. Defaults to \code{TRUE}}

\item{interactive}{should put R object in global environment? Defaults to \code{TRUE}}

\item{stn.orf_find.all_starts}{\code{orf_find.all_starts} parameter for the \code{SaTAnn} function}

\item{stn.orf_find.nostarts}{\code{orf_find.nostarts} parameter for the \code{SaTAnn} function}

\item{stn.orf_find.start_sel_cutoff}{\code{orf_find.start_sel_cutoff} parameter for the \code{SaTAnn} function}

\item{stn.orf_find.start_sel_cutoff_ave}{\code{orf_find.start_sel_cutoff_ave} parameter for the \code{SaTAnn} functio}

\item{stn.orf_find.cutoff_fr_ave}{\code{orf_find.cutoff_fr_ave} parameter for the \code{SaTAnn} function}

\item{stn.orf_quant.cutoff_cums}{\code{orf_quant.cutoff_cums} parameter for the \code{SaTAnn} function}

\item{stn.orf_quant.cutoff_pct}{\code{orf_quant.cutoff_pct} parameter for the \code{SaTAnn} function}

\item{stn.orf_quant.cutoff_P_sites}{\code{orf_quant.cutoff_P_sites} parameter for the \code{SaTAnn} function}
}
\value{
A set of output files containing transcript coordinates, exonic coordinates and annotation for each ORF, including optional GTF and protein fasta files.\cr\cr
The description for each list object is as follows:\cr\cr
\code{tmp_SaTAnn_results}: (Optional) RData object file containing the entire set of results for each genomic region.\cr
\code{final_SaTAnn_results}: RData object file containing the final SaTAnn results, see \code{SaTAnn}.\cr
\code{Protein_sequences.fasta}: (Optional) Fasta file containing the set of translated proteins .\cr
\code{Detected_ORFs.gtf}: GTF file containing coordinates of the detected ORFs.\cr\cr
In addition, new columns are added in the ORFs_tx file:\cr\cr
\code{TrP_pM}: (Beta) multitaper spectral coefficient of the P_sites track for each ORF, summing up to a million.\cr
\code{TrP_pN}: (Beta) multitaper spectral coefficient of the P_sites track for each ORF, divided by ORF length.\cr
\code{TrP_pNpM}: (Beta) multitaper spectral coefficient of the P_sites track for each ORF, divided by ORF length and summing up to a million (akin to TPM).\cr
\code{P_sites_pM}: number of P_sites for each ORF, summing up to a million.\cr
\code{P_sites_pN}: number of P_sites for each ORF, divided by ORF length.\cr
\code{P_sites_pNpM}: number of P_sites for each ORF, divided by ORF length and summing up to a million (akin to TPM).\cr
}
\description{
This wrapper function runs the entire SaTAnn pipeline
}
\details{
A set of transcripts, together with genome sequence and Ribo-signal are analyzed to extract translated ORFs
}
\seealso{
\code{\link{prepare_annotation_files}}, \code{\link{load_annotation}}, \code{\link{SaTAnn}}
}
\author{
Lorenzo Calviello, \email{calviello.l.bio@gmail.com}
}
\keyword{SaTAnn}
